refactor: defines the meta tags and changes the name of the character traits component
