+++ src/domain/RouteMetaTags.interface.ts
@@ -1,5 +1,8 @@
+type MetaNames = 'description' | 'keywords' | 'og:title' | 'og:description' | 'og:image' | 'addsearch-custom-field';
+
 export interface RouteMetaTags {
-    name?: string;
+    name?: MetaNames;
     content: string;
     property?: string;
+    'data-type'?: string;
 }
