+++ src/domain/character/UseColorTags.interface.ts
@@ -1,0 +1,6 @@
+import { ComputedRef } from 'vue';
+
+export interface UseColorTags {
+    colorTagSpecie: ComputedRef<string>;
+    colorTagStatus: ComputedRef<string>;
+}
