+++ src/infrastructure/router/middleware/removeAllMetaTags.ts
@@ -1,4 +1,7 @@
-export const removeAllMetaTags = () => {
+/**
+ * Removes all meta tags from the head of the document.
+ */
+export const removeAllMetaTags = (): void => {
     const { head } = document;
 
     const metaTags = head.getElementsByTagName('meta') as HTMLCollectionOf<HTMLMetaElement>;
